requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.2.4
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.58.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
yapf==0.31.0
zipp==3.4.1
